accessible/tests/mochitest/test_elm_table.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=483573
-->
<head>
  <title>File Input Control tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      var accTree = {
        role: ROLE_TABLE, // table
        children: [
          {
            role: ROLE_TEXT_CONTAINER, // thead
            children: [
              {
                role: ROLE_ROW,
                children: [
                  {
                    role: ROLE_COLUMNHEADER
                  },
                  {
                    role: ROLE_COLUMNHEADER
                  }
                ]
              }
            ]
          },
          {
            role: ROLE_TEXT_CONTAINER, // tbody
            children: [
              {
                role: ROLE_ROW,
                children: [
                  {
                    role: ROLE_CELL
                  },
                  {
                    role: ROLE_CELL
                  }
                ]
              }
            ]
          },
          {
            role: ROLE_TEXT_CONTAINER, // tfoot
            children: [
              {
                role: ROLE_ROW,
                children: [
                  {
                    role: ROLE_COLUMNHEADER
                  },
                  {
                    role: ROLE_COLUMNHEADER
                  }
                ]
              }
            ]
          }
        ]
      };

      if (LINUX)
        todo(false, "No tests on linux because of different hierarchies.");
      else
        testAccessibleTree("table", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="create accessibles for HTML tr"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=493695">Mozilla Bug 493695</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <table id="table">
    <thead>
      <tr>
        <th>col1</th><th>col2</th>
      </tr>
    </thead>
    <tbody>
      <tr>
        <td>cell1</td><td>cell2</td>
      </tr>
    </tbody>
    <tfoot>
      <tr>
        <th>col1</th><th>col2</th>
      </tr>
    </tfoot>
  </table>
</body>
</html>
